move {Number,Boolean}Param to core/parameters
