+++ src/client-builtins/mod.ts
@@ -2,9 +2,7 @@ import "./meshViewer3d.tsx";
 
 // moderate
 
-export * from "./BooleanParam.tsx";
 export * from "./EdgesGeometry.ts";
-export * from "./NumberParam.tsx";
 export * from "./Viewer3d.tsx";
 export * from "./axes.ts";
 export * from "./colors.ts";
