+++ src/client/BooleanParam.tsx
@@ -1,5 +1,5 @@
 // @style "./stylus/BooleanParam.styl"
-import { BooleanParam } from "../builtins/mod.ts";
+import { BooleanParam } from "../core/mod.ts";
 import React from "../deps/react.ts";
 import { observer } from "../deps/rhobo.ts";
 import { NameDesc, registerParameter } from "../client/mod.ts";
