+++ src/client/NumberParam.tsx
@@ -1,5 +1,5 @@
 // @style "./stylus/NumberParam.styl"
-import { NumberParam } from "../builtins/mod.ts";
+import { NumberParam } from "../core/mod.ts";
 import React from "../deps/react.ts";
 import { observer, useComputed, useObservable } from "../deps/rhobo.ts";
 import { NameDesc, registerParameter } from "../client/mod.ts";
