+++ src/core/parameters/mod.ts
@@ -1,4 +1,6 @@
 // moderate
 
+export * from "./BooleanParam.ts";
+export * from "./NumberParam.ts";
 export * from "./ObjectParam.ts";
 export * from "./Parameter.ts";
